+++ cpp-opencv-app/testy/Sklejanie-test-wysajności/CMakeLists.txt
