+++ cpp-opencv-app/testy/Sklejanie-test-wysajności/README.md
@@ -1,0 +1,7 @@
+## C++ OpenCV App
+
+1.  mkdir build
+2.  cd build
+3.  cmake ..
+4.  make
+5.  ./opencv-app
