+++ cpp-opencv-app/testy/Sklejanie_obrazu/CMakeLists.txt
