+++ cpp-opencv-app/testy/Sklejanie_obrazu/main.cpp
