+++ cpp-opencv-app/testy/UART_test_pakowania_danych_LINUX/main.cpp
