+++ cpp-opencv-app/wersja_końcowa/CMakeLists.txt
