+++ cpp-opencv-app/wersja_końcowa/README.md
