+++ cpp-opencv-app/wersja_końcowa/include/GUI.cpp
