+++ cpp-opencv-app/wersja_końcowa/include/camera_GPU.h
