+++ cpp-opencv-app/wersja_końcowa/main.cpp
